+++ NEWS.md
@@ -1,6 +1,6 @@
 # tinycodet 0.5.3 (Under Development)
 * **Behaviour change:** the atomic typecasting functions now preserve names, dimensions, and dimnames (instead of all attributes), to be more in line with most of base 'R'.
-* **Behaviour change:** `pkgs %installed in% lib.loc` will now return `FALSE` when core 'R' `pkgs` are not in `lib.loc`, in contrast to `find.package()`.
+* **Behaviour change:** `pkgs %installed in% lib.loc` will now return `NA` for "packages" that are part of core 'R' (i.e. 'base', 'stats', etc.).
 * **Behaviour change:** All decimal truth testing operators now always give `NA` when `Inf` is compared with `Inf`, or when `-Inf` is compared to `-Inf`.
 * Streamlined the internal code of the decimal truth testing operators.
 * Added `%s><%` and `%s<>%` as aliases for `%sget%` and `%strim%`, respectively.